# docs/swagger.yaml
definitions:
  handler.CreateReq:
    properties:
      content:
        type: string
      title:
        type: string
    required:
    - content
    - title
    type: object
  handler.LoginReq:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  handler.RegisterReq:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
info:
  contact: {}
paths:
  /api/articles:
    post:
      consumes:
      - application/json
      description: 需要登录，创建新文章，自动关联当前用户
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: 文章信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 创建文章
      tags:
      - 文章
  /api/login:
    post:
      consumes:
      - application/json
      description: 用户名密码登录，返回 JWT token
      parameters:
      - description: 登录凭证
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 用户名或密码错误
          schema:
            additionalProperties: true
            type: object
      summary: 用户登录
      tags:
      - 用户
  /api/register:
    post:
      consumes:
      - application/json
      description: 注册新用户，返回成功信息（密码会自动哈希）
      parameters:
      - description: 注册信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.RegisterReq'
      produces:
      - application/json
      responses:
        "200":
          description: 注册成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties: true
            type: object
      summary: 用户注册
      tags:
      - 用户
swagger: "2.0"